tests/pyproject.toml
[tool.poetry]
name = "bff_cli"
version = "2.2.10"
description = "This service is designed to support command line interface"
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi-utils = "^0.7"
uvicorn = "0.12.3"
gunicorn = "20.0.4"
uvloop = "0.20.0"
httptools = "^0.6"
SQLAlchemy = "1.4.31"
python-dotenv = "0.19.1"
asyncpg = "0.29.0"
opentelemetry-exporter-jaeger = "1.6.2"
opentelemetry-instrumentation = "0.26b1"
opentelemetry-instrumentation-fastapi = "0.26b1"
opentelemetry-instrumentation-sqlalchemy = "0.26b1"
opentelemetry-instrumentation-httpx = "0.26b1"
opentelemetry-instrumentation-asyncpg = "0.26b1"
fastapi-health = "^0.4.0"
fastapi = "^0.90"
pilot-platform-common = "0.3.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "^3.0.0"
pytest-asyncio = "0.17.2"
pytest-httpx="^0.21.0"
pytest-random-order = "1.1.1"
pytest-mock = "3.7.0"
faker = "28.4.1"
sqlalchemy-utils = "0.38.2"
uvicorn = "0.12.3"
gunicorn = "20.0.4"
requests = "2.24.0"
python-multipart = "0.0.5"
aiofiles = "0.6.0"
redis = "3.5.3"
async-asgi-testclient = "1.4.9"
testcontainers = "3.4.2"

[tool.pytest.ini_options]
addopts = "-v --random-order"
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["setuptools", "poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
